Move ReportCreator, add CreationMode
Fix UI report generation
