+++ src/back/Forum/Report/CreationMode.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace KeepersTeam\Webtlo\Forum\Report;
+
+/** Тип формирования отчёта. */
+enum CreationMode
+{
+    case CRON;
+    case UI;
+}
